+++ src/api/quest/repository.ts
@@ -1,5 +1,8 @@
-import { apiClient } from "../../../pkg/api/client/apiClient";
+import { apiClient } from "../../pkg/api/client/apiClient";
 import {
+  CreateRequest,
+  CreateResponse,
+  DeleteRequest,
   FinishRequest,
   FinishResponse,
   ForceFinishRequest,
@@ -7,10 +10,15 @@ import {
   ListResponse,
   StartRequest,
   StartResponse,
+  UpdateRequest,
+  UpdateResponse,
 } from "./types";
 
 export interface QuestRepository {
   list: () => Promise<ListResponse>;
+  create: (params: CreateRequest) => Promise<CreateResponse>;
+  destroy: (params: DeleteRequest) => Promise<void>;
+  update: (params: UpdateRequest) => Promise<UpdateResponse>;
   start: (params: StartRequest) => Promise<StartResponse>;
   finish: (params: FinishRequest) => Promise<FinishResponse>;
   forceFinish: (params: ForceFinishRequest) => Promise<ForceFinishResponse>;
@@ -18,6 +26,20 @@ export interface QuestRepository {
 
 const list: QuestRepository["list"] = async () => {
   const response = await apiClient.get("/quests");
+  return response;
+};
+
+const create: QuestRepository["create"] = async (params: CreateRequest) => {
+  const response = await apiClient.post("/quests", params);
+  return response;
+};
+
+const destroy: QuestRepository["destroy"] = async (params: DeleteRequest) => {
+  await apiClient.destroy(`/quests/${params.id}`);
+};
+
+const update: QuestRepository["update"] = async (params: UpdateRequest) => {
+  const response = await apiClient.patch(`/quests/${params.id}`, params);
   return response;
 };
 
@@ -38,6 +60,9 @@ const forceFinish: QuestRepository["finish"] = async (params: ForceFinishRequest
 
 export const questRepository: QuestRepository = {
   list,
+  create,
+  destroy,
+  update,
   start,
   finish,
   forceFinish,
