--- src/features/quests/api/mock/data.ts
@@ -1,243 +1,0 @@
-// import { formatDate, now } from "../../../pkg/util/dayjs";
-// import { Quest } from "../api/model";
-
-// export const dataMock: Quest[] = [
-//   {
-//     id: "0",
-//     title: "英語単語の学習",
-//     description: "前回の続きから50単語やる",
-//     startsAt: `${formatDate(now())} 09:00:00`,
-//     startedAt: `${formatDate(now())} 09:00:00`,
-//     minutes: 30,
-//     tag_id: "タグ名1",
-//     difficulty: "EASY",
-//     state: "INACTIVE",
-//     isSucceeded: false,
-//     startDate: "",
-//     end_date: "",
-//     dates: ["月", "火", "水"],
-//     weeklyFrequency: 3,
-//     weeklyCompletionCount: 0,
-//   },
-//   {
-//     id: "1",
-//     title: "英語音読",
-//     description: "音読7回を行う",
-//     startsAt: `${formatDate(now())} 09:50:00`,
-//     startedAt: `${formatDate(now())} 09:40:00`,
-//     minutes: 30,
-//     tag_id: "タグ名1",
-//     difficulty: "EASY",
-//     state: "INACTIVE",
-//     isSucceeded: false,
-//     startDate: "",
-//     end_date: "",
-//     dates: ["月", "火", "水"],
-//     weeklyFrequency: 3,
-//     weeklyCompletionCount: 0,
-//   },
-//   {
-//     id: "3",
-//     title: "クエスト名3",
-//     description: "クエスト説明3です。",
-//     startsAt: `${formatDate(now())} 13:00:00`,
-//     startedAt: `${formatDate(now())} 12:45:00`,
-//     minutes: 25,
-//     tag_id: "タグ名3",
-//     difficulty: "MID",
-//     state: "INACTIVE",
-//     isSucceeded: false,
-//     startDate: "",
-//     end_date: "",
-//     dates: ["月", "水", "金"],
-//     weeklyFrequency: 3,
-//     weeklyCompletionCount: 0,
-//   },
-//   {
-//     id: "2",
-//     title: "クエスト名2",
-//     description: "クエスト説明2です。",
-//     startsAt: `${formatDate(now())} 12:00:00`,
-//     startedAt: "NOT_STARTED_YET",
-//     minutes: 25,
-//     tag_id: "タグ名2",
-//     difficulty: "MID",
-//     state: "INACTIVE",
-//     isSucceeded: false,
-//     startDate: "",
-//     end_date: "",
-//     dates: ["月", "水", "金"],
-//     weeklyFrequency: 3,
-//     weeklyCompletionCount: 0,
-//   },
-//   {
-//     id: "4",
-//     title: "クエスト名4",
-//     description: "クエスト説明4です。",
-//     startsAt: `${formatDate(now())} 14:00:00`,
-//     startedAt: `${formatDate(now())} 13:45:00`,
-//     minutes: 25,
-//     tag_id: "タグ名4",
-//     difficulty: "MID",
-//     state: "INACTIVE"e,
-//     isSucceeded: false,
-//     startDate: "",
-//     end_date: "",
-//     dates: ["月", "水", "金"],
-//     weeklyFrequency: 3,
-//     weeklyCompletionCount: 0,
-//   },
-//   {
-//     id: "6",
-//     title: "クエスト名6",
-//     description: "クエスト説明6です。",
-//     startsAt: `${formatDate(now())} 16:00:00`,
-//     startedAt: "NOT_STARTED_YET",
-//     minutes: 25,
-//     tag_id: "タグ名6",
-//     difficulty: "MID",
-//     state: "INACTIVE"e,
-//     isSucceeded: false,
-//     startDate: "",
-//     end_date: "",
-//     dates: ["月", "水", "金"],
-//     weeklyFrequency: 3,
-//     weeklyCompletionCount: 0,
-//   },
-//   {
-//     id: "5",
-//     title: "クエスト名5",
-//     description: "クエスト説明5です。",
-//     startsAt: `${formatDate(now())} 15:00:00`,
-//     startedAt: "NOT_STARTED_YET",
-//     minutes: 25,
-//     tag_id: "タグ名5",
-//     difficulty: "MID",
-//     state: "INACTIVE"e,
-//     isSucceeded: false,
-//     startDate: "",
-//     end_date: "",
-//     dates: ["月", "水", "金"],
-//     weeklyFrequency: 3,
-//     weeklyCompletionCount: 0,
-//   },
-//   {
-//     id: "7",
-//     title: "クエスト名7",
-//     description: "クエスト説明7です。",
-//     startsAt: `${formatDate(now())} 17:00:00`,
-//     startedAt: "NOT_STARTED_YET",
-//     minutes: 25,
-//     tag_id: "タグ名7",
-//     difficulty: "MID",
-//     state: "INACTIVE"e,
-//     isSucceeded: false,
-//     startDate: "",
-//     end_date: "",
-//     dates: ["月", "水", "金"],
-//     weeklyFrequency: 3,
-//     weeklyCompletionCount: 0,
-//   },
-//   {
-//     id: "9",
-//     title: "クエスト名9",
-//     description: "クエスト説明9です。",
-//     startsAt: `${formatDate(now())} 19:00:00`,
-//     startedAt: "NOT_STARTED_YET",
-//     minutes: 25,
-//     tag_id: "タグ名9",
-//     difficulty: "MID",
-//     state: "INACTIVE"e,
-//     isSucceeded: false,
-//     startDate: "",
-//     end_date: "",
-//     dates: ["月", "水", "金"],
-//     weeklyFrequency: 3,
-//     weeklyCompletionCount: 0,
-//   },
-//   {
-//     id: "8",
-//     title: "クエスト名8",
-//     description: "クエスト説明8です。",
-//     startsAt: `${formatDate(now())} 18:00:00`,
-//     startedAt: "NOT_STARTED_YET",
-//     minutes: 25,
-//     tag_id: "タグ名8",
-//     difficulty: "MID",
-//     state: "INACTIVE"e,
-//     isSucceeded: false,
-//     startDate: "",
-//     end_date: "",
-//     dates: ["月", "水", "金"],
-//     weeklyFrequency: 3,
-//     weeklyCompletionCount: 0,
-//   },
-//   {
-//     id: "10",
-//     title: "クエスト名10",
-//     description: "クエスト説明10です。",
-//     startsAt: `${formatDate(now())} 20:50:00`,
-//     startedAt: "NOT_STARTED_YET",
-//     minutes: 25,
-//     tag_id: "タグ名10",
-//     difficulty: "MID",
-//     state: "INACTIVE"e,
-//     isSucceeded: false,
-//     startDate: "",
-//     end_date: "",
-//     dates: ["月", "水", "金"],
-//     weeklyFrequency: 3,
-//     weeklyCompletionCount: 0,
-//   },
-//   {
-//     id: "11",
-//     title: "クエスト名10",
-//     description: "クエスト説明10です。",
-//     startsAt: `${formatDate(now())} 21:00:00`,
-//     startedAt: `NOT_STARTED_YET`,
-//     minutes: 5,
-//     tag_id: "タグ名10",
-//     difficulty: "MID",
-//     state: "INACTIVE"e,
-//     isSucceeded: false,
-//     startDate: "",
-//     end_date: "",
-//     dates: ["月", "水", "金"],
-//     weeklyFrequency: 3,
-//     weeklyCompletionCount: 0,
-//   },
-//   {
-//     id: "12",
-//     title: "クエスト名10",
-//     description: "クエスト説明10です。",
-//     startsAt: `${formatDate(now())} 21:45:00`,
-//     startedAt: "NOT_STARTED_YET",
-//     minutes: 25,
-//     tag_id: "タグ名10",
-//     difficulty: "MID",
-//     state: "INACTIVE"e,
-//     isSucceeded: false,
-//     startDate: "",
-//     end_date: "",
-//     dates: ["月", "水", "金"],
-//     weeklyFrequency: 3,
-//     weeklyCompletionCount: 0,
-//   },
-//   {
-//     id: "13",
-//     title: "クエスト名10",
-//     description: "クエスト説明10です。",
-//     startsAt: `${formatDate(now())} 23:20:00`,
-//     startedAt: `NOT_STARTED_YET`,
-//     minutes: 3,
-//     tag_id: "タグ名10",
-//     difficulty: "MID",
-//     state: "INACTIVE"e,
-//     isSucceeded: false,
-//     startDate: "",
-//     end_date: "",
-//     dates: ["月", "水", "金"],
-//     weeklyFrequency: 3,
-//     weeklyCompletionCount: 0,
-//   },
-// ];
